fix: Tweak kit for exchange integration
