+++ src/widgets/Modal/index.tsx
@@ -1,3 +1,4 @@
 export { default as Modal } from "./Modal";
 export { default as ModalProvider } from "./ModalContext";
 export { default as useModal } from "./useModal";
+export type { ModalProps } from "./types";
