+++ src/widgets/Modal/types.ts
@@ -1,3 +1,7 @@
-export type ModalTheme = {
+export interface ModalTheme {
   background: string;
-};
+}
+
+export interface ModalProps {
+  onDismiss: () => void;
+}
